+++ test/src/validators/validators_test.dart
@@ -60,6 +60,32 @@ void main() {
       expect(control.valid, true);
     });
 
+    test('FormGroup valid if minLength valid', () {
+      // Given: a valid control
+      final form = FormGroup({
+        'name': FormControl(
+          value: 'Reactive Forms',
+          validators: [Validators.minLength(6)],
+        ),
+      });
+
+      // Expect: control is valid
+      expect(form.valid, true);
+    });
+
+    test('FormGroup invalid if minLength invalid', () {
+      // Given: an invalid control
+      final form = FormGroup({
+        'name': FormControl(
+          value: 'Forms',
+          validators: [Validators.minLength(6)],
+        ),
+      });
+
+      // Expect: control is invalid
+      expect(form.valid, false);
+    });
+
     test('FormArray invalid if minLength invalid', () {
       // Given: an invalid array
       final array = FormArray([
@@ -134,6 +160,29 @@ void main() {
 
       // Expect: control is valid
       expect(control.valid, true);
+    });
+
+    test('FormGroup invalid if maxLength invalid', () {
+      final form = FormGroup({
+        'name': FormControl(
+          value: 'Hello Reactive Forms',
+          validators: [Validators.maxLength(10)],
+        ),
+      });
+
+      expect(form.invalid, true);
+      expect(form.hasError(ValidationMessage.maxLength, 'name'), true);
+    });
+
+    test('FormGroup is valid if maxLength valid', () {
+      final form = FormGroup({
+        'name': FormControl(
+          value: 'Reactive',
+          validators: [Validators.maxLength(10)],
+        ),
+      });
+
+      expect(form.valid, true);
     });
 
     test('FormControl of list valid if maxLength is valid', () {
@@ -871,6 +920,90 @@ void main() {
       expect(() => CompareValidator('', null, CompareOption.equal),
           throwsAssertionError);
       expect(() => CompareValidator('', '', null), throwsAssertionError);
+    });
+  });
+
+  group('Contain validator tests', () {
+    test('Compare a List of Strings with a String of an email address (invalid)', () {
+      final control = FormControl<List<String>>(
+          value: ['john@email.com'],
+          validators: [Validators.contains(['1','3'])],
+      );
+
+      // Expect: control invalid
+      expect(control.valid, false);
+    });
+
+    test('Compare a List of Strings with a String of all numbers (invalid)', () {
+      final control = FormControl<List<String>>(
+          value: ['0123456789'],
+          validators: [Validators.contains(['1','3'])],
+      );
+
+      // Expect: control invalid
+      expect(control.valid, false);
+    });
+
+    test('Compare a List of Strings with another that contains all of them (valid)', () {
+      final control = FormControl<List<String>>(
+          value: ['1','2','3','4'],
+          validators: [Validators.contains(['1','3'])],
+      );
+
+      // Expect: control valid
+      expect(control.valid, true);
+    });
+
+    test('Compare a list of Stings with a List of numbers (invalid)', () {
+      final control = FormControl<List<int>>(
+          value: [1,2,3,4],
+          validators: [Validators.contains(['1','3'])],
+      );
+
+      // Expect: control invalid
+      expect(control.valid, false);
+    });
+
+    test('Compare a list of numbers with another that contains all of them (valid)', () {
+      final control = FormControl<List<int>>(
+          value: [1,2,3,4],
+          validators: [Validators.contains([1,3])],
+      );
+
+      // Expect: control invalid
+      expect(control.valid, true);
+    });
+
+    test('Compare a list of numbers with another that contains a part of them (invalid)', () {
+      final control = FormControl<List<int>>(
+          value: [1,2,3,4],
+          validators: [Validators.contains([1,3,5])],
+      );
+
+      // Expect: control invalid
+      expect(control.valid, false);
+    });
+
+    test('Compare a list of emails with another that contains numbers (invalid)', () {
+      final control = FormArray<String>([
+        FormControl<String>(value: 'john@email.com'),
+        FormControl<String>(value: 'stark@email.com'),
+        FormControl<String>(value: 'bob@email.com'),
+      ], validators: [Validators.contains(['doe@email.com'])]);
+
+      // Expect: control invalid
+      expect(control.valid, false);
+    });
+
+    test('Compare a list of numbers with another that contains all of them (valid)', () {
+      final control = FormArray<int>([
+        FormControl<int>(value: 1),
+        FormControl<int>(value: 2),
+        FormControl<int>(value: 3),
+      ], validators: [Validators.contains([1,3])]);
+
+      // Expect: control invalid
+      expect(control.valid, true);
     });
   });
 }